+++ explorer/views.py
@@ -12,28 +12,26 @@ from django.forms.models import model_to_dict
 from explorer.models import Query, QueryLog
 from explorer import app_settings
 from explorer.forms import QueryForm
-from explorer.utils import url_get_rows, url_get_query_id, url_get_log_id, schema_info, url_get_params, safe_admin_login_prompt, build_download_response, build_stream_response
+from explorer.utils import url_get_rows, url_get_query_id, url_get_log_id, schema_info, url_get_params, safe_admin_login_prompt, build_download_response, build_stream_response, user_can_see_query
 
 from collections import Counter
+from functools import wraps
 import re
 
 
 def view_permission(f):
+    @wraps(f)
     def wrap(request, *args, **kwargs):
-        if not app_settings.EXPLORER_PERMISSION_VIEW(request.user) and not user_can_see_query(request, kwargs):
+        if not app_settings.EXPLORER_PERMISSION_VIEW(request.user)\
+                and not user_can_see_query(request, kwargs)\
+                and not (app_settings.EXPLORER_TOKEN_AUTH_ENABLED and request.META.get('X-API-TOKEN') == app_settings.EXPLORER_TOKEN):
             return safe_admin_login_prompt(request)
         return f(request, *args, **kwargs)
     return wrap
 
 
-def user_can_see_query(request, kwargs):
-    if not request.user.is_anonymous() and 'query_id' in kwargs:
-        allowed_queries = app_settings.EXPLORER_GET_USER_QUERY_VIEWS().get(request.user.id, [])
-        return int(kwargs['query_id']) in allowed_queries
-    return False
-
-
 def change_permission(f):
+    @wraps(f)
     def wrap(request, *args, **kwargs):
         if not app_settings.EXPLORER_PERMISSION_CHANGE(request.user):
             return safe_admin_login_prompt(request)
